fixed order of sockets emitting and listening for events for a more seamless experience, cleaned up css files
