+++ frontend/src/components/Questions/Questions.styles.js
@@ -1,0 +1,28 @@
+const styles = {
+  sx: {
+    QuestionContainer: {
+      display: "flex",
+      flexDirection: "column",
+      justifyContent: "center",
+      justifyItems: "center",
+      alignContent: "center",
+      alignItems: "center",
+      width: "300px",
+    },
+    SelectedAnswer: {
+      width: "20%",
+      fontSize: "10px",
+      backgroundColor: "cadetBlue",
+      margin: "2%",
+      border: "2px solid black",
+    },
+    NonselectedAnswer: {
+      width: "20%",
+      fontSize: "10px",
+      backgroundColor: "cadetBlue",
+      margin: "2%",
+    },
+  },
+};
+
+export default styles;
